#44 [rename] constant 파일에 도메인명을 접두어로 추가
